Including Record-Replay functionality
Modifying the source-code to include OPTIONAL record-replay functionality offered by VCRPy. If a cassette is specified, the request is stored in a cassette, which can then be replayed later. This commit is currently known to break requests made using SSL. Once this is fixed, I shall go ahead and submit another pull request with updated test cases.